+++ src/backend/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod collection;
 pub(crate) mod config;
 pub mod definition;
+pub(crate) mod extract;
 pub(crate) mod release;
 pub mod rename;
 pub mod semantic;