+++ src/lib.rs
@@ -1,8 +1,8 @@
 pub use mlg::{
-    CheckResult, CleanResult, DebugResult, ExportResult, FormatResult, InitResult, LspResult,
-    ReleaseResult, VersionResult, ViewResult, WhteRbtObjResult, check, check_diagnostics_report,
-    check_diagnostics_schema, clean, debug, export, format, init, lsp, release, version, view,
-    whte_rbt_obj,
+    CheckResult, CleanResult, DebugResult, ExportResult, ExtractResult, FormatResult, InitResult,
+    LspResult, ReleaseResult, ReportResult, VersionResult, ViewResult, WhteRbtObjResult, check,
+    check_diagnostics_report, check_diagnostics_schema, clean, debug, export, extract, format,
+    init, lsp, release, report, version, view, whte_rbt_obj,
 };
 pub mod backend;
 pub mod cli;